\name{arithmeticRefsMedoids}
\docType{data}
\alias{arithmeticRefsMedoids}
\title{Object each reference file is assigned as it is loaded in or read.}
\description{
  arithmeticRefsMedoids is the name of the object each reference is
  assigned to after it is loaded or read.
}
\usage{arithmeticRefsMedoids}
\format{A list of two character vectors, the first containing the name
  of each SNP and the second containing the arithmetic medoid value for each SNP.}
\source{Ancestry Mapper from various datasets. See: http://bit.ly/1OUstDP}
\references{
 Various
}
\keyword{datasets}
