\name{MinMaxFreq}
\docType{data}
\alias{MinMaxFreq}
\title{Allele Variants for Demo SNPs According to dbSNP.}
\description{
  This data set contains the major and minor alleles for the 1000 SNPs
  used in the demo data according to the strand orientation used on dbSNP.
}
\usage{MinMaxFreq}
\format{A matrix containing 1000 observations of two variables.}
\source{dbSNP.}
\references{
  "ftp://ftp.ncbi.nlm.nih.gov/snp/organisms/human_9606/VCF/00-All.vcf.gz"
}
\keyword{datasets}
